requirements.txt
google-genai>=1.0.0
Pillow>=10.0.0
tqdm>=4.66.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
